pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.10.1",
]
